Move pagination methods to ApplicationRecord
