+++ app/controllers/api/v1/merchants/items_controller.rb
@@ -1,0 +1,5 @@
+class Api::V1::Merchants::ItemsController < ApplicationController
+  def index
+
+  end
+end
